Add authorization and work with the database
Added routes for working with users and transactions. Connected mongodb,
wrote schemes for the user and transactions.
Transactions defined two endpoints for receiving one and several
transactions. Defined user registration and user acquisition
authorization endpoints for the user. Added a class for working with a
token. Described work with the user. Rewrote requests to work with the
server